Crop_Disease_Prediction/requirements.txt
-f https://download.pytorch.org/whl/cpu/torch-1.10.1%2Bcpu-cp37-cp37m-linux_x86_64.whl
-f https://files.pythonhosted.org/packages/91/a6/d852cacd9b2444adfc3f821cabdc7c0d442b184139a720f631cbb3e91360/Pillow-6.2.2-pp373-pypy36_pp73-win32.whl
torch==1.10.1+cpu
Flask==2.1.1
flask_cors
gunicorn
numpy==1.21.4
pillow==6.2.2
pandas
config
requests==2.27.1
scikit-learn==1.0.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
torchvision==0.2.2
tqdm==4.19.9
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
yarg==0.1.9